chore: breaking out into function
