+++ packages/cli/src/lib/new/factories/backendPlugin.ts
@@ -20,7 +20,7 @@ import camelCase from 'lodash/camelCase';
 import { paths } from '../../paths';
 import { addCodeownersEntry, getCodeownersFilePath } from '../../codeowners';
 import { CreateContext, createFactory } from '../types';
-import { addPackageDependency, Task } from '../../tasks';
+import { addPackageDependency, addToBackend, Task } from '../../tasks';
 import { ownerPrompt, pluginIdPrompt } from './common/prompts';
 import { executePluginPackageTemplate } from './common/tasks';
 import { resolvePackageName } from './common/util';
@@ -79,29 +79,9 @@ export const backendPlugin = createFactory<Options>({
         );
       });
 
-      await Task.forItem('backend', 'adding plugin', async () => {
-        const backendFilePath = paths.resolveTargetRoot(
-          'packages/backend/src/index.ts',
-        );
-        if (!(await fs.pathExists(backendFilePath))) {
-          return;
-        }
-
-        const content = await fs.readFile(backendFilePath, 'utf8');
-        const lines = content.split('\n');
-        const backendAddLine = `backend.add(import('${name}'));`;
-
-        const backendStartIndex = lines.findIndex(line =>
-          line.match(/backend.start/),
-        );
-
-        if (backendStartIndex !== -1) {
-          const [indentation] = lines[backendStartIndex].match(/^\s*/)!;
-          lines.splice(backendStartIndex, 0, `${indentation}${backendAddLine}`);
-
-          const newContent = lines.join('\n');
-          await fs.writeFile(backendFilePath, newContent, 'utf8');
-        }
+      await addToBackend(name, {
+        defaultVersion: ctx.defaultVersion,
+        type: 'plugin',
       });
     }
