+++ packages/cli/src/lib/tasks.ts
@@ -196,3 +196,35 @@ export async function addPackageDependency(
     throw new Error(`Failed to add package dependencies, ${error}`);
   }
 }
+
+export async function addToBackend(
+  name: string,
+  ctx: { defaultVersion: string; type: 'plugin' | 'module' },
+) {
+  if (await fs.pathExists(paths.resolveTargetRoot('packages/backend'))) {
+    await Task.forItem('backend', `adding ${ctx.type}`, async () => {
+      const backendFilePath = paths.resolveTargetRoot(
+        'packages/backend/src/index.ts',
+      );
+      if (!(await fs.pathExists(backendFilePath))) {
+        return;
+      }
+
+      const content = await fs.readFile(backendFilePath, 'utf8');
+      const lines = content.split('\n');
+      const backendAddLine = `backend.add(import('${name}'));`;
+
+      const backendStartIndex = lines.findIndex(line =>
+        line.match(/backend.start/),
+      );
+
+      if (backendStartIndex !== -1) {
+        const [indentation] = lines[backendStartIndex].match(/^\s*/)!;
+        lines.splice(backendStartIndex, 0, `${indentation}${backendAddLine}`);
+
+        const newContent = lines.join('\n');
+        await fs.writeFile(backendFilePath, newContent, 'utf8');
+      }
+    });
+  }
+}
